Move rateLimit to own file
