+++ package.json
@@ -29,6 +29,7 @@
   "devDependencies": {
     "@types/cron": "^2.0.1",
     "@types/lodash": "^4.14.195",
+    "@types/uuid": "^9.0.2",
     "@typescript-eslint/eslint-plugin": "^5.59.11",
     "@typescript-eslint/parser": "^5.59.11",
     "auto": "^10.44.0",
